main <== develop (배포 환경 테스트 후 변경 사항 반영)

+++ common/src/main/java/com/daangndaangn/common/jwt/Jwt.java
@@ -57,7 +57,7 @@ public final class Jwt {
         builder.withIssuedAt(now);
 
         if (expirySeconds > 0) {
-            builder.withExpiresAt(new Date(now.getTime() + expirySeconds * 1_000L));
+            builder.withExpiresAt(new Date(now.getTime() + (long)expirySeconds * 1_000L));
         }
 
         builder.withClaim("id", claims.id);
